% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gamma.score.test}
\alias{gamma.score.test}
\title{gamma.score.test(): function to compute p-value of score test given model fit, step size (dh), order (accuracy, 1, 2 or 4) and data set}
\usage{
\method{gamma}{score.test}(fit, dh, accuracy, data)
}
\description{
gamma.score.test(): function to compute p-value of score test given model fit, step size (dh), order (accuracy, 1, 2 or 4) and data set
}
