% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gamma.simulator}
\alias{gamma.simulator}
\title{gamma.simulator(): function to simulate data from our model for a specified shape parameter (k), screening interval, with/without background risk (include.h)
note:
  - parameters go in order (background risk, lambda_1, lambda_2, prevalent)
  - this code is simplified to not include covariates}
\usage{
\method{gamma}{simulator}(n, params, k, interval = 3, include.h = T)
}
\description{
gamma.simulator(): function to simulate data from our model for a specified shape parameter (k), screening interval, with/without background risk (include.h)
note:
  - parameters go in order (background risk, lambda_1, lambda_2, prevalent)
  - this code is simplified to not include covariates
}
