% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{model.fit}
\alias{model.fit}
\title{Fit the prevalence-incidence-cure model}
\usage{
model.fit(
  l1_x = c(),
  l2_x = c(),
  pi_x = c(),
  data,
  epsilon = 1e-08,
  short.epsilon = 0.1,
  short.iter = 10,
  short.runs = 20,
  silent = T,
  init = NULL,
  include.h = T,
  two.step.h = T,
  include.priors = T,
  fixed.h = NULL,
  intercept.prog = T,
  intercept.clear = T,
  intercept.prev = T
)
}
\arguments{
\item{l1_x}{A vector containing the names of covariates used in the progression rate parameter (must match column name(s) in the input data), can be blank}

\item{l2_x}{A vector containing the names of covariates used in the clearance rate parameter (must match column name(s) in the input data), can be blank}

\item{pi_x}{A vector containing the names of covariates used in the parameter (probability of prevalent disease) (must match column name(s) in the input data), can be blank}

\item{data}{Data used to fit the model containing columns for each term in l1_x, l2_x and pi_x. The first three columns must be (1)
left interval, (2) right interval and (3) z indicator for prevalent/incident disease.The data must contain observed prevalent and incident cases, however some cases may be unknown (`NA`).}

\item{epsilon}{Convergence criteria for the change in log-likelihood value used for stopping the EM algorithm. Defaults to 1e-08.}

\item{short.epsilon}{Convergence criteria used in the 'Short' EM algorithm to determine initial values. Defaults to 0.1.}

\item{silent}{Indicator variable for whether to print out each iteration or not. Default value set to FALSE.}

\item{num.runs}{Number of runs of the 'Short' EM algorithm used to determine initial values for the EM algorithm. Defaults to 30.}
}
\value{
The output is a list containing the following elements:
  initial.values - initial values determined by the Short EM algorithm process
  theta.hat - optimum parameter values estimated by the EM algorithm
  num.iterations - number of iterations until the EM algorithm converged
  log.likelihood - value of the log.likelihood at the optimum parameter values
  hess - hessian matrix
  std.dev - standard deviation of parameter estimates
  summary - data frame with estimate, std.dev, and 95\% CI for each parameter (useful for data set comparisons)
}
\description{
Fit the prevalence-incidence-cure model
}
