% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fit.R
\name{PICmodel.fit}
\alias{PICmodel.fit}
\title{Fit the prevalence-incidence-clearance model}
\usage{
PICmodel.fit(
  data,
  prog_model = "prog ~ 1",
  prev_model = "prev ~ 1",
  epsilon = 1e-06,
  short.epsilon = 0.1,
  short.iter = 10,
  short.runs = 20,
  silent = T,
  init = NULL,
  max.iter = 1000,
  include.h = T,
  starting.h = -12,
  include.priors = T,
  prior.type = "cauchy",
  fixed.h = NULL,
  intercept.prog = T,
  intercept.clear = T,
  intercept.prev = T
)
}
\arguments{
\item{data}{Data used to fit the model containing columns for each term in l1_x and pi_x. The first three columns must be: \itemize{
\item left interval giving the last time a patient was seen without the disease
\item right interval giving the first time a patient was seen with the disease (can be Inf for right-censored cases)
\item z indicator for prevalent/incident disease. The data must contain observed prevalent (z=1) and incident (z=0) cases, however some cases may be set to unknown (i.e., z=`NA`) for instance if there were no visits between baseline and disease detection.
}}

\item{prog_model}{A string in the format "prog ~ x1 + x2 + ...", where x1, x2 are the names of covariates used in the progression rate parameter (must match column name(s) in the input data). Defaults to "prog ~ 1", which is an intercept only model (i.e., without covariates).}

\item{prev_model}{A string in the format "prev ~ x1 + x2 + ...", where x1, x2 are the names of covariates used in the parameter (probability of prevalent disease) (must match column name(s) in the input data). Defaults to "prev ~ 1", which is an intercept only model (i.e., without covariates).}

\item{epsilon}{Convergence criteria for the change in log-likelihood value used for stopping the EM algorithm. Defaults to 1e-08.}

\item{short.epsilon}{Convergence criteria used in the 'Short' EM algorithm to determine initial values. Defaults to 0.1.}

\item{short.iter}{Number of max iterations used in one run of the 'Short' EM algorithm used to determine initial values for the EM algorithm. Defaults to 10}

\item{short.runs}{Number of runs of the 'Short' EM algorithm used to determine initial values for the EM algorithm. Defaults to 20}

\item{silent}{Indicator variable for whether to print results to console during iterations. Defaults to FALSE. Results print in order: (1) background risk (if used), (2) progression model, (3) clearance model, (4) prevalence model, and (5) log-likelihood. Parameters are on the log-scale, except for prevalence which is on the logit scale.}

\item{init}{Option to supply initial values instead of starting from random values in the model fit procedure. Defaults to NULL.}

\item{max.iter}{Number of maximum iterations used in the final run of the EM algorithm. Defaults to 1000.}

\item{include.h}{Indicator variable for whether to include background risk in the model or not. Defaults to TRUE.}

\item{starting.h}{Option to supply a value for an initial value of background risk in the model. Defaults to -12 (on the log-scale).}

\item{include.priors}{Indicator variable for whether to include priors in the estimation procedure. Defaults to TRUE.}

\item{prior.type}{Choice of prior, either "Cauchy" or "t4" (for a t-distribution prior with 4 degrees of freedom) prior. Defaults to the weakly informative Cauchy prior.}

\item{fixed.h}{Option to supply a fixed value of background risk, in which case it is treated as fixed and not included in parameter estimation and summary statistics. Needs to be on the log scale, e.g., background risk of 0.01 will use $log(0.01) = -4.6$ as input. Defaults to NULL.}

\item{intercept.prog}{Indicator variable for whether a general intercept is estimated for the progression parameter. Used when adding country/study specific dummy variables instead. Defaults to TRUE.}

\item{intercept.clear}{Indicator variable for whether a general intercept is estimated for the clearance parameter. Used when adding country/study specific dummy variables instead. Defaults to TRUE.}

\item{intercept.prev}{Indicator variable for whether a general intercept is estimated for the prevalence parameter. Used when adding country/study specific dummy variables instead. Defaults to TRUE.}
}
\value{
The output is a list containing the following elements:
\itemize{
\item `model`: list containing names of covariates used for progression and prevalence parameter
\item `initial.values`: either the initial values determined by the Short EM algorithm process or the supplied initial values
\item `fixed.h`: the fixed value of background risk if supplied, otherwise NULL
\item `theta.hat`: optimum parameter values estimated by the EM algorithm
\item `num.iterations`: number of iterations until the EM algorithm converged
\item `log.likelihood`: value of the log.likelihood at the maximum likelihood estimates
\item `hess`: hessian matrix at the maximum likelihood estimates
\item `grad`: value of the gradient vector at the the maximum likelihood estimates
\item `std.dev`: standard deviation of parameter estimates
\item `convergence`: indicator variable for whether the algorithm converged or reached maximum number of iterations. If FALSE, try with higher max.iter, smaller epsilon or different starting values.
\item `summary`: data frame with estimate, std.dev, and 95\% CI for each parameter (useful for data set comparisons). Parameters that start with `g` (i.e., `g0`, `g1`...) are for the progression model, those that start with `w` (i.e., `w0`) are for the clearance model, and those start with `p` (i.e., `p0`, `p1`...) are for the prevalence model. Progression and clearance parameters are on the log scale, prevalence parameters are on logit scale.
}
}
\description{
This function fits our Prevalence-Incidence-Clearance (PIC) mixture model to interval-censored screening data and obtains parameter estimates.
It is possible for the user to select the covariates that will be used for each parameter.
}
\examples{
fit <- PICmodel.fit(sim.dat)
fit$summary
}
\author{
Kelsi Kroon, Hans Bogaards, Hans Berkhof
}
