% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_simulator.R
\name{PICmodel.simulator}
\alias{PICmodel.simulator}
\title{Screening data simulator}
\usage{
PICmodel.simulator(
  n,
  prog_model = "prog ~ 1",
  prev_model = "prev ~ 1",
  params,
  show_prob = 0.9,
  interval = 3,
  include.h = T,
  covar_settings = c(age.min = 30, age.max = 70, cyt.prob = 0.4, hpv16.prob = 0.3)
)
}
\arguments{
\item{n}{Number of women in the simulated data set.}

\item{prog_model}{A string in the format "prog ~ x1 + x2 + ...", where x1, x2 are the names of covariates used in the progression rate parameter (must match column name(s) in the input data). Defaults to "prog ~ 1", which is an intercept only model (i.e., without covariates).}

\item{prev_model}{A string in the format "prev ~ x1 + x2 + ...", where x1, x2 are the names of covariates used in the parameter (probability of prevalent disease) (must match column name(s) in the input data). Defaults to "prev ~ 1", which is an intercept only model (i.e., without covariates).}

\item{params}{Numerical vector the parameter values to be used in the data simulation (first value is background risk, then l1, l2, pi), including the values of the covariates.}

\item{show_prob}{A value representing the probability of a woman showing up for the screening visit. Defaults to 0.9.}

\item{interval}{A value representing the interval between screening rounds (in years). Defaults to 3.}

\item{include.h}{Indicator variable for whether to background risk in the simulation procedure. Defaults to TRUE.}

\item{covar_settings}{A vector of settings used to generate covariate values for age, abnormal cytology and HPV16 genotype. They must be supplied in the order age minimum, age maximum, abnormal cytology probability and HPV16 probability. Age is generated uniformly between `age.min` and `age.max`, with default values of 30 and 70 years respectively. Abnormal cytology is generated from a Bernoulli distribution with probability `cyt.prop`, with default probability of 0.4. HPV16 is also generated from a Bernoulli distribution with probability `hpv16.prob', which defaults to 0.3.}
}
\value{
A data frame containing the left and right interval of CIN2+ detection, the indicator of prevalent disease, age (continuous), HPV genotype (HPV16 or other, 1=HPV16), and cytology result (normal or abnormal, 1=abnormal).
}
\description{
Simulates screening data with user-specified parameters loosely based on the example of cervical cancer screening data. Useful for validating that the model is able to recover the true parameter values. Currently it is only possible to simulate
data with the baseline covariates age (continuous between 30 and 70), HPV genotype (HPV16 positive or negative), and cytology (normal/ abnormal).
}
\examples{
sim.df <- PICmodel.simulator(1000, prog_model = "prog ~ hpv",  c(-5, -1.6, 1, -1.2, 0.25), show_prob = 0.9, interval=3, include.h=T)
head(sim.df)
}
\author{
Kelsi Kroon, Hans Bogaards, Hans Berkhof
}
