% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gamma.R
\name{gamma.score.test}
\alias{gamma.score.test}
\title{Gamma Score test}
\usage{
\method{gamma}{score.test}(fit, dh, accuracy, data)
}
\value{
The output is a list containing the following elements:
\itemize{
\item grad: value of the gradient at the shape parameter equal to 1 and the rest of the parameters at their maximum likelihood estimates.
\item inv.hess: value of the inverse hessian at the shape parameter equal to 1 and the rest of the parameters at their maximum likelihood estimates.
\item hess: value of the hessian at the shape parameter equal to 1 and the rest of the parameters at their maximum likelihood estimates.
\item loglik: value of the log-likelihood at the shape parameter equal to 1 and the rest of the parameters at their maximum likelihood estimates.
\item p.val: p-value of the score test for whether Cause 1 follows a Gamma rather than Exponential distribution (tests whether shape parameter equals 1 or not).
}
}
\description{
Function to compute p-value of the score test given model fit, step size (dh), order (accuracy, 1, 2 or 4) and data.
}
