% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gamma.R
\name{gamma.simulator}
\alias{gamma.simulator}
\title{Gamma Simulator}
\usage{
\method{gamma}{simulator}(n, params, k, interval = 3, include.h = T)
}
\arguments{
\item{n}{Number of subjects requried in simulated data set}

\item{params}{Parameter values for simple model without covariates. Note that parameters go in order (1) background risk, (2) lambda_1, (3), lambda_2, and (4) prevalent).}

\item{k}{Shape parameter for gamma distribution}

\item{interval}{Screening interval (in years). Defaults to 3.}

\item{include.h}{Indicator variable for whether to include background risk or not. Defaults to TRUE.}
}
\description{
Function to simulate data for a variant of our model that uses the Gamma distribution for Cause 1 (i.e., progression) in the competing risks framework used for incident disease. Note this code is simplified to not include covariates
}
