% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fit.R
\name{model.fit}
\alias{model.fit}
\title{Fit the prevalence-incidence-cure model}
\usage{
model.fit(
  l1_x = c(),
  l2_x = c(),
  pi_x = c(),
  data,
  epsilon = 1e-08,
  short.epsilon = 0.1,
  short.iter = 10,
  short.runs = 20,
  silent = T,
  init = NULL,
  include.h = T,
  two.step.h = T,
  include.priors = T,
  fixed.h = NULL,
  intercept.prog = T,
  intercept.clear = T,
  intercept.prev = T
)
}
\arguments{
\item{l1_x}{A vector containing the names of covariates used in the progression rate parameter (must match column name(s) in the input data), can be blank}

\item{l2_x}{A vector containing the names of covariates used in the clearance rate parameter (must match column name(s) in the input data), can be blank}

\item{pi_x}{A vector containing the names of covariates used in the parameter (probability of prevalent disease) (must match column name(s) in the input data), can be blank}

\item{data}{Data used to fit the model containing columns for each term in l1_x, l2_x and pi_x. The first three columns must be (1)
left interval, (2) right interval and (3) z indicator for prevalent/incident disease.The data must contain observed prevalent and incident cases, however some cases may be unknown (`NA`).}

\item{epsilon}{Convergence criteria for the change in log-likelihood value used for stopping the EM algorithm. Defaults to 1e-08.}

\item{short.epsilon}{Convergence criteria used in the 'Short' EM algorithm to determine initial values. Defaults to 0.1.}

\item{short.iter}{Number of max iterations used in one run of the 'Short' EM algorithm used to determine initial values for the EM algorithm. Defaults to 10}

\item{short.runs}{Number of runs of the 'Short' EM algorithm used to determine initial values for the EM algorithm. Defaults to 20}

\item{silent}{Indicator variable for whether to print out each iteration or not. Default value set to FALSE.}

\item{init}{Option to supply initial values instead of starting from random values in the model fit procedure. Defaults to NULL.}

\item{include.h}{Indicator variable for whether to include background risk in the model or not. Defaults to TRUE.}

\item{two.step.h}{Indicator variable for whether to perform a two-step procedure to obtain initial values when background risk is included. Defaults to TRUE.}

\item{include.priors}{Indicator variable for whether to incldue weakly informative Cauchy priors in the estimation procedure. Defaults to TRUE.}

\item{fixed.h}{Option to supply a fixed value of background risk, in which case it is treated as fixed and not included in parameter estimation and summary statistics. Needs to be on the log scale, e.g., background risk of 0.01 will use $log(0.01) = 4.6$ as input. Defaults to NULL.}

\item{intercept.prog}{Defaults to TRUE.}

\item{intercept.clear}{Defaults to TRUE.}

\item{intercept.prev}{Defaults to TRUE.}
}
\value{
The output is a list containing the following elements:
\itemize{
\item model: list containing names of covariates used for each parameter
\item initial.values: either the initial values determined by the Short EM algorithm process or the supplied initial values
\item fixed.h: the fixed value of background risk if supplied, otherwise NULL
\item theta.hat: optimum parameter values estimated by the EM algorithm
\item num.iterations: number of iterations until the EM algorithm converged
\item log.likelihood: value of the log.likelihood at the maximum likelihood estimates
\item hess: hessian matrix at the maximum likelihood estimates
\item grad: value of the gradient vector at the the maximum likelihood estimates
\item std.dev: standard deviation of parameter estimates
\item summary: data frame with estimate, std.dev, and 95\% CI for each parameter (useful for data set comparisons)
}
}
\description{
This function fits our Prevalence-Incidence-Cure (PIC) mixture model to interval-censored screening data and obtains parameter estimates.
It is possible for the user to select the covariates that will be used for each parameter.
}
