% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_predict.R
\name{PICmodel.predict}
\alias{PICmodel.predict}
\title{Make predictions from the prevalence-incidence-cure model}
\usage{
PICmodel.predict(data, time.points, fit, calc.CI = F)
}
\arguments{
\item{data}{Data set of covariates from which to make the predictions. Must contain columns that correspond to covariates used in `fit`.}

\item{time.points}{Numeric vector of time points used to make cumulative risk predictions}

\item{fit}{Parameter estimates for the model to be used (output from model.fit)}

\item{calc.CI}{Indicator variable for whether confidence intervals for the cumulative risk should be calculated. Defaults to FALSE.}
}
\value{
The output is a list for each unique combination of covariates used in the model containing the time points, the cumulative risk estimate along with the upper and lower 95% confidence intervals.
}
\description{
Make predictions from the prevalence-incidence-cure model
}
\author{
Kelsi Kroon, Hans Bogaards, Hans Berkhof
}
