% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_simulator.R
\name{PICmodel.simulator}
\alias{PICmodel.simulator}
\title{Screening data simulator}
\usage{
PICmodel.simulator(
  n,
  l1_x,
  l2_x,
  pi_x,
  params,
  show_prob = 0.9,
  interval = 3,
  include.h = T
)
}
\arguments{
\item{n}{Number of women in the simulated data set.}

\item{l1_x}{A vector containing the names of covariates used in the \ifelse{html}{\out{\eqn{\lambda}<sub>1</sub>}}{ \eqn{\lambda_1}} (progression rate) parameter. Options are "age", "HPV16" and "cytology".}

\item{l2_x}{A vector containing the names of covariates used in the \eqn{\lambda_2} (clearance rate) parameter. Options are "age", "HPV16" and "cytology".}

\item{pi_x}{A vector containing the names of covariates used in the \eqn{\pi} parameter (probability of prevalent disease). Options are "age", "HPV16" and "cytology".}

\item{params}{Numerical vector the parameter values to be used in the data simulation (first value is background risk, then l1, l2, pi)}

\item{show_prob}{A value representing the probability of a woman showing up for the screening visit. Defaults to 0.9.}

\item{interval}{A value representing the interval between screening rounds (in years). Defaults to 3.}

\item{include.h}{Indicator variable for whether to background risk in the simulation procedure. Defaults to TRUE.}
}
\value{
A data frame containing the left and right interval of CIN2+ detection, the indicator of prevalent disease, age (continuous), HPV genotype (HPV16 or other, 1=HPV16), and cytology result (normal or abnormal, 1=abnormal).
}
\description{
Simulates screening data with user-specified parameters loosely based on the example of cervical cancer screening data. Useful for validating that the model is able to recover the true parameter values. Currently it is only possible to simulate
data with the baseline covariates age (continuous between 30 and 70), HPV genotype (HPV16 positive or negative), and cytology (normal/ abnormal).
}
\examples{
sim.df <- PICmodel.simulator(1000, c("hpv"), c(), c(),  c(-5, -1.6, 1, -1.2, 0.25), show_prob = 0.9, interval=3, include.h=T)
head(sim.df)
}
\author{
Kelsi Kroon, Hans Bogaards, Hans Berkhof
}
